refactor: import문 쿼리  파라미터 제거
