+++ frontend/src/components/specific/modal/singlePhotoModal/SinglePhotoModal.tsx
@@ -1,4 +1,4 @@
-import defaultImage from '../../../../@assets/images/default-forgather-image.png?w=600&format=webp&srcset';
+import defaultImage from '../../../../@assets/images/default-forgather-image.png';
 import { createImageErrorHandler } from '../../../../utils/createImageErrorHandler';
 
 import Modal from '../../../@common/modal/Modal';
